+++ src/backend/serial/mod.rs
@@ -21,13 +21,19 @@
 //! `32bit` since identifiers can't start with letters, and the backends
 //! do use `u32`/`u64`, so this seems like a least-bad option.
 
-#[cfg(not(any(feature = "u32_backend", feature = "u64_backend", feature = "avx2_backend")))]
+#[cfg(not(any(feature = "u32_backend", feature = "u64_backend")))]
 compile_error!(
     "no curve25519-dalek backend cargo feature enabled! \
-     please enable one of: u32_backend, u64_backend, avx2_backend"
+     please enable one of: u32_backend, u64_backend"
 );
 
-pub mod serial;
+#[cfg(feature = "u32_backend")]
+pub mod u32;
 
-#[cfg(all(feature = "avx2_backend", target_feature = "avx2"))]
-pub mod vector;
+#[cfg(feature = "u64_backend")]
+pub mod u64;
+
+pub mod curve_models;
+
+pub mod scalar_mul;
+
