+++ src/backend/serial/u32/scalar.rs
