+++ src/backend/serial/u64/scalar.rs
