+++ src/backend/vector/avx2/mod.rs
@@ -27,4 +27,3 @@ pub(crate) mod edwards;
 
 pub(crate) mod constants;
 
-pub(crate) mod scalar_mul;
