+++ src/backend/vector/mod.rs
@@ -21,13 +21,12 @@
 //! `32bit` since identifiers can't start with letters, and the backends
 //! do use `u32`/`u64`, so this seems like a least-bad option.
 
-#[cfg(not(any(feature = "u32_backend", feature = "u64_backend", feature = "avx2_backend")))]
-compile_error!(
-    "no curve25519-dalek backend cargo feature enabled! \
-     please enable one of: u32_backend, u64_backend, avx2_backend"
-);
-
-pub mod serial;
-
 #[cfg(all(feature = "avx2_backend", target_feature = "avx2"))]
-pub mod vector;
+pub mod avx2;
+#[cfg(all(feature = "avx2_backend", target_feature = "avx2"))]
+pub(crate) use self::avx2::{
+    constants::BASEPOINT_ODD_LOOKUP_TABLE, edwards::CachedPoint, edwards::ExtendedPoint,
+};
+
+pub mod scalar_mul;
+
